+++ source/gameengine/Converter/BL_ConvertActuators.cpp
@@ -919,7 +919,7 @@ void BL_ConvertActuators(const char *maggiename,
         if (_2dfilter->text) {
           char *buf;
           // this is some blender specific code
-          int buf_len_dummy;
+          size_t buf_len_dummy;
           buf = txt_to_buf(_2dfilter->text, &buf_len_dummy);
           if (buf) {
             tmp->SetShaderText(buf);
